requirements.txt
pandas
scikit-learn
numpy
nltk
python-docx
pdfplumber
openpyxl
fitz
